+++ app/src/main/java/com/websarva/wings/android/reversingvulnapp/ui/fragment/top/RecyclerViewHolder.kt
@@ -1,9 +1,9 @@
-package com.websarva.wings.android.vulnmemory_aes.ui.fragment.top
+package com.websarva.wings.android.reversingvulnapp.ui.fragment.top
 
 import android.view.View
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.websarva.wings.android.vulnmemory_aes.R
+import com.websarva.wings.android.reversingvulnapp.R
 
 class RecyclerViewHolder(var view: View): RecyclerView.ViewHolder(view) {
     val title: TextView = view.findViewById(R.id.tvTitle)
